move compare_rots_sym to utils
